+++ src/kaocha/plugin/notifier.clj
@@ -11,8 +11,8 @@
             [clojure.java.io :as io]
             [slingshot.slingshot :refer [throw+]])
   (:import [java.nio.file Files]
-           [java.io IOException] 
-           [java.awt SystemTray TrayIcon  TrayIcon$MessageType Toolkit])) 
+           [java.io IOException]
+           [java.awt SystemTray TrayIcon TrayIcon$MessageType Toolkit]))
 
 ;; special thanks for terminal-notify stuff to
 ;; https://github.com/glittershark/midje-notifier/blob/master/src/midje/notifier.clj
@@ -20,11 +20,10 @@
 (defn exists? [program]
   (let [cmd (if (platform/on-windows?)
               "where.exe" "which" )]
-    (try 
+    (try
       (= 0 (:exit (sh cmd program)))
       (catch IOException e  ;in the unlikely event where.exe or which isn't available
         (output/warn (format "Unable to determine whether '%s' exists. Notifications may not work." program)) ))))
-
 
 (defn detect-command []
   (cond
@@ -64,31 +63,30 @@
            (io/copy (io/make-input-stream resource {}) file)
            (str file)))))))
 
-
 (def tray-icon
   "Creates a system tray icon."
   (memoize
-    (fn [icon-path]
-      (let [tray-icon (-> (Toolkit/getDefaultToolkit) 
-                          (.getImage icon-path)
-                          (TrayIcon. "Kaocha Notification"))]
-        (doto (SystemTray/getSystemTray)
-          (.add tray-icon))
-        tray-icon))))
+   (fn [icon-path]
+     (let [^java.awt.Toolkit toolkit (Toolkit/getDefaultToolkit)
+           tray-icon (-> toolkit
+                         (.getImage ^String icon-path)
+                         (TrayIcon. "Kaocha Notification"))]
+       (doto (SystemTray/getSystemTray)
+         (.add tray-icon))
+       tray-icon))))
 
-(defn send-tray-notification 
+(defn send-tray-notification
   "Use Java's built-in functionality to display a notification.
 
   Not preferred over shelling out because the built-in notification sometimes
   looks out of place, and isn't consistently available on Linux."
   [result]
-  (try 
+  (try
     (let [icon (tray-icon "kaocha/clojure_logo.png")
           urgency (if (result/failed? result) TrayIcon$MessageType/ERROR TrayIcon$MessageType/INFO) ]
       (.displayMessage icon (title result) (message result) urgency))
     (catch java.awt.HeadlessException e
       (output/warn (str "Notification not shown because system is headless. AWT error: " e)) )))
-
 
 (defn expand-command
   "Takes a command string including replacement patterns, and a map of